refactor - WIP
