+++ examples/sampleAtomCommand.css
@@ -1,4 +1,4 @@
-.category {
+.atom-category {
     color: #fff;
     margin-right: 6px;
     border-radius: 2px;
