+++ examples/sampleAtomCommand.js
@@ -1,9 +1,10 @@
 import React from "react";
+import "./sampleAtomCommand.css";
 
 export default function sampleAtomCommand(suggestion) {
   const { name, highlight, shortcut } = suggestion;
   return (
-    <div className="item">
+    <div className="atom-item">
       {highlight ? (
         <span dangerouslySetInnerHTML={{ __html: highlight }} />
       ) : (
