+++ themes/chrome-theme.js
